+++ CHANGELOG.md
@@ -9,8 +9,15 @@
   - **HÁ BUGS**
 - Alterações visuais de ponteiro
 
-## Versão 1.2 -
+## Versão 1.2 - 09/12/2025
 - Volume vai de 0% a 100%
 - Novo sistema de conquistas com 3 níveis
   - 17 conquistas para liberar
 - Otimização de código
+
+## Versão 1.3 - 10/12/2025
+- Novas 12 conquistas
+- Novos textos informativos nas configurações
+- Correção de ver a produção de cookies/seg em cada upgrade
+- Reformulação do sistema de salvar
+- Otimizações de código